# Parallel processing repairing playlist