@touwys 2023-09-24
> Expected result:
> 
>     * should do process the repair process faster, but is strongly depended on hardware (number of CPU cores).
> 
>     * results in playlist are updated out of sequence

_Tested with a c 13-year-old Windows 7 x64 PC with a 4-core Intel Core i7-3770K CPU. These observations are not absolutely observed:_

_Numerous playlists were submitted for test-repair. Those tested, contain any number of tracks from 19 up to 631._

### Result:

1. There is a striking ⚡ gain of speed noticeable during the first round of playlist repair — i.e. at finding the **exact matches**. 

2. As it were, these exact matches were also found and updated out-of-sequence. 

3. The speed gain on the second round of repairs for each of the playlists — i.e. to find the **closest matches** — was not as clearly visible as in the first round. Superficially, it appears to be not on par with the search for the exact matches, but, presumably, that is due to any differences inherent between the two processes? For instance, are the searches for closest matches carried out out-of-sequence, too? Even so, even if the speed gain here is not as clearly established, the processing happens fast enough to be wholly acceptable.

@Borewit: Great job — once again. 🏆