@Borewit 2023-09-24
I am looking at the closest matches algorithm, looks like Jeremy optimized the algorithm for you PC @touwys: 

https://github.com/Borewit/listFix/blob/9a9c96a74d795579e2d8ea0e4859f6b8df1dab4e/src/main/java/listfix/model/playlists/PlaylistEntry.java#L148-L150

🤣