@Borewit 2023-09-24
Despite the funny comment, I removed that optimization and replaced that with parallel processing as well. Looking forward to hear if the _find-closest-match_ algorithm runs faster as well @touwys.