@touwys 2023-09-24
> [Build run (#6290622796, attempt #1) artifacts](https://github.com/Borewit/listFix/actions/runs/6290622796/attempts/1#artifacts)

Is it ready to try out? I can only attend to it tomorrow.

***

> Despite the funny comment, I removed that optimization and replaced that with parallel processing as well. Looking forward to hear if the find-closest-match algorithm runs faster as well.

I hope that it is as noticeable a difference as with finding the exact matches. What really counts though, is the accuracy of the matches, rather than the speed by which they are delivered. 

What I have seen so far, is that the current search algorithm is actually quite accurate for finding matching tracks where the broken ones form part of various artist compilations. Obviously, it's got more data available to process, in the file name of the broken track. So, one could probably retain that strength, and focus on massaging the algorithm where it tries to locate the tracks with very little data available in the filename.