@Borewit 2023-09-24
It's good thing you didn't test yet. I was not the smartest location to apply parallel processing, I adjust the processing keeping both parallel processing on Jeremy's memory optimization for now.

For me the closest match runs very fast, so hard to notice a big difference.

In this PR nothing changed from a functional point of view. So the accuracy has not been changed. On my Intel(R) Core(TM) i7-7800X CPU @ 3.50GHz, 6 cores (but 12 logical cores):

**Before this PR:**
Repaired playlist in 9491 ms.
Resolved closest matches in 433 ms.

**After this PR:**
Repaired playlist in 3512 ms.
Resolved closest matches in 391 ms.