@touwys 2023-09-25
> **Before this PR:** Repaired playlist in 9491 ms. Resolved closest matches in 433 ms.
> 
> **After this PR:** Repaired playlist in 3512 ms. Resolved closest matches in 391 ms.

@Borewit :

The advance in your results is clearly very impressive.

Even if the huge difference in hardware makes a straightforward comparison of the results quite impossible, I easily recognised the improvement in my results. You've certainly done something special in this.

To what extent do other variables such as, for e.g. the number of tracks in the media database, parsing of the track information by the search algorithm, etc. play a role in advancing the search speed?

***

> **It's good thing you didn't test yet.** I was not the smartest location to apply parallel processing, I adjust the processing keeping both parallel processing on Jeremy's memory optimization for now.

Which build should I then download to use next, or should I wait for another? Please indicate.

***