@Borewit 2023-09-25
> To what extent do other variables such as, for e.g. the number of tracks in the media database, parsing of the track information by the search algorithm, etc. play a role in advancing the search speed?

I expect no significant difference. With a background job which takes at least a few seconds, the advantage of parallel processing will likely by higher then a relative very short job, as result of a more optimal spread and the overhead being relatively smaller.  

Gaining small advantages resulting in more complex code, I am not much in favor of.  In line with [_"Premature Optimization is the  Root of All Evil"_](https://effectiviology.com/premature-optimization/). The code to optimization done for the "users with an ancient computer" is probably an example of that. It is almost asking for trouble.  

If you can please re-test [the latest build](https://github.com/Borewit/listFix/pull/203#issuecomment-1732627098)